Introduce RelativeCube abstraction for tracking orientation
In order to allow for color-neutrality (and just rotations in general),
we have to maintain the state of the cube "as the solver sees it" in
order to know which facelets actually represent locations of interest
(the front-right F2L slot, for example).

Previously, an Array of all prior rotation moves was maintained and
applied to a brand-new cube every time we wanted to resolve relative
locations. We want to do that quite often, so a secondary cube now
maintains the rotation history and can be asked to resolve locations
on-demand. Some care was taken to ensure that `RelativeCube`s do not
have `RelativeCube`s of their own, which would blow the stack real fast.